mcp-servers/mcp-common/pyproject.toml
[project]
name = "mcp-common"
version = "0.1.0"
description = "Shared helpers for self-service-agent MCP servers (headers, tracing)"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "mcp[cli]>=1.23",
    "opentelemetry-exporter-otlp-proto-http>=1.37.0",
    "self-service-agent-shared-models",
    "tracing-config",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/mcp_common"]

[tool.uv.sources]
self-service-agent-shared-models = { path = "../../shared-models" }
tracing-config = { path = "../../tracing-config" }

[tool.mypy]
python_version = "3.12"
mypy_path = "../../tracing-config/src:../../shared-models/src"

[[tool.mypy.overrides]]
module = ["opentelemetry.*", "shared_models.*"]
follow_imports = "skip"
